refactor: move files to new package view.map
