+++ src/main/scala/view/map/MapView.scala
@@ -1,4 +1,4 @@
-package view
+package view.map
 
 import javax.swing.JComponent
 import scala.swing.*
