+++ src/test/scala/view/map/CitiesLoaderTest.scala
@@ -1,4 +1,4 @@
-package view
+package view.map
 
 import org.scalatest.flatspec.AnyFlatSpec
 import org.scalatest.matchers.should.Matchers
